Restructuring Apply ViewModels
