+++ src/SFA.DAS.AssessorService.Web.Staff/SFA.DAS.AssessorService.Web.Staff.csproj
@@ -128,4 +128,8 @@
     </Content>
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="ViewModels\Apply\Applications\" />
+  </ItemGroup>
+
 </Project>
