+++ src/SFA.DAS.AssessorService.Web.Staff/Validators/FinancialApplicationViewModelValidator.cs
@@ -1,7 +1,7 @@
 using FluentValidation;
 using SFA.DAS.AssessorService.ApplyTypes;
 using SFA.DAS.AssessorService.Web.Staff.Controllers.Apply;
-using SFA.DAS.AssessorService.Web.Staff.ViewModels;
+using SFA.DAS.AssessorService.Web.Staff.ViewModels.Apply.Financial;
 
 namespace SFA.DAS.AssessorService.Web.Staff.Validators
 {
