+++ src/SFA.DAS.AssessorService.Web.Staff/ViewModels/Apply/Financial/NewFinancialApplicationViewModel.cs
@@ -1,6 +1,6 @@
 using System;
 
-namespace SFA.DAS.AssessorService.Web.Staff.ViewModels
+namespace SFA.DAS.AssessorService.Web.Staff.ViewModels.Apply.Financial
 {
     public class NewFinancialApplicationViewModel
     {
