+++ src/SFA.DAS.AssessorService.Web.Staff/Views/Apply/Financial/Application.cshtml
@@ -1,5 +1,5 @@
 @using SFA.DAS.AssessorService.ApplyTypes
-@model SFA.DAS.AssessorService.Web.Staff.ViewModels.FinancialApplicationViewModel
+@model SFA.DAS.AssessorService.Web.Staff.ViewModels.Apply.Financial.FinancialApplicationViewModel
 
 @{
     ViewBag.Title = "title";
