+++ src/SFA.DAS.AssessorService.Web.Staff/Views/Apply/Financial/NewApplications.cshtml
@@ -1,4 +1,4 @@
-@model List<SFA.DAS.AssessorService.Web.Staff.ViewModels.NewFinancialApplicationViewModel>
+@model List<SFA.DAS.AssessorService.Web.Staff.ViewModels.Apply.Financial.NewFinancialApplicationViewModel >
 
 @{
     ViewBag.Title = "New applications";
