+++ src/SFA.DAS.AssessorService.Web.Staff/Views/Apply/Financial/PreviousApplication.cshtml
@@ -1,6 +1,6 @@
 @using Microsoft.EntityFrameworkCore.Internal
 @using Microsoft.EntityFrameworkCore.Query.Internal
-@model SFA.DAS.AssessorService.Web.Staff.ViewModels.FinancialApplicationViewModel
+@model SFA.DAS.AssessorService.Web.Staff.ViewModels.Apply.Financial.FinancialApplicationViewModel
 
 @{
     ViewBag.Title = "title";
