+++ src/SFA.DAS.AssessorService.Web.Staff/Views/Apply/Financial/PreviousApplications.cshtml
@@ -1,4 +1,4 @@
-@model List<SFA.DAS.AssessorService.Web.Staff.ViewModels.PreviousFinancialApplicationViewModel>
+@model List<SFA.DAS.AssessorService.Web.Staff.ViewModels.Apply.Financial.PreviousFinancialApplicationViewModel>
 
 @{
     ViewBag.Title = "Previous applications";
